% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_moonshot.R
\name{KimiCodeAnthropicProvider}
\alias{KimiCodeAnthropicProvider}
\title{Kimi Code Anthropic Provider Class}
\description{
Anthropic-compatible provider wrapper for Kimi Code.
}
\keyword{internal}
\section{Super class}{
\code{\link[aisdk:AnthropicProvider]{aisdk::AnthropicProvider}} -> \code{KimiCodeAnthropicProvider}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KimiCodeAnthropicProvider-language_model}{\code{KimiCodeAnthropicProvider$language_model()}}
\item \href{#method-KimiCodeAnthropicProvider-clone}{\code{KimiCodeAnthropicProvider$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="AnthropicProvider" data-id="enable_caching"><a href='../../aisdk/html/AnthropicProvider.html#method-AnthropicProvider-enable_caching'><code>aisdk::AnthropicProvider$enable_caching()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="AnthropicProvider" data-id="initialize"><a href='../../aisdk/html/AnthropicProvider.html#method-AnthropicProvider-initialize'><code>aisdk::AnthropicProvider$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KimiCodeAnthropicProvider-language_model"></a>}}
\if{latex}{\out{\hypertarget{method-KimiCodeAnthropicProvider-language_model}{}}}
\subsection{Method \code{language_model()}}{
Create a Kimi Code Anthropic-compatible language model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KimiCodeAnthropicProvider$language_model(model_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_id}}{The model ID. Defaults to \verb{kimi-for-coding}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An AnthropicLanguageModel object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KimiCodeAnthropicProvider-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KimiCodeAnthropicProvider-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KimiCodeAnthropicProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
