% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_moonshot.R
\name{MoonshotLanguageModel}
\alias{MoonshotLanguageModel}
\title{Moonshot Language Model Class}
\description{
Language model implementation for Moonshot / Kimi Chat Completions APIs.
}
\keyword{internal}
\section{Super classes}{
\code{\link[aisdk:LanguageModelV1]{aisdk::LanguageModelV1}} -> \code{\link[aisdk:OpenAILanguageModel]{aisdk::OpenAILanguageModel}} -> \code{MoonshotLanguageModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MoonshotLanguageModel-new}{\code{MoonshotLanguageModel$new()}}
\item \href{#method-MoonshotLanguageModel-build_payload}{\code{MoonshotLanguageModel$build_payload()}}
\item \href{#method-MoonshotLanguageModel-build_stream_payload}{\code{MoonshotLanguageModel$build_stream_payload()}}
\item \href{#method-MoonshotLanguageModel-parse_response}{\code{MoonshotLanguageModel$parse_response()}}
\item \href{#method-MoonshotLanguageModel-clone}{\code{MoonshotLanguageModel$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="LanguageModelV1" data-id="generate"><a href='../../aisdk/html/LanguageModelV1.html#method-LanguageModelV1-generate'><code>aisdk::LanguageModelV1$generate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="LanguageModelV1" data-id="has_capability"><a href='../../aisdk/html/LanguageModelV1.html#method-LanguageModelV1-has_capability'><code>aisdk::LanguageModelV1$has_capability()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="LanguageModelV1" data-id="stream"><a href='../../aisdk/html/LanguageModelV1.html#method-LanguageModelV1-stream'><code>aisdk::LanguageModelV1$stream()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="do_generate"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-do_generate'><code>aisdk::OpenAILanguageModel$do_generate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="do_stream"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-do_stream'><code>aisdk::OpenAILanguageModel$do_stream()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="execute_request"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-execute_request'><code>aisdk::OpenAILanguageModel$execute_request()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="format_tool_result"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-format_tool_result'><code>aisdk::OpenAILanguageModel$format_tool_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="get_config"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-get_config'><code>aisdk::OpenAILanguageModel$get_config()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="get_history_format"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-get_history_format'><code>aisdk::OpenAILanguageModel$get_history_format()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoonshotLanguageModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-MoonshotLanguageModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Moonshot language model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoonshotLanguageModel$new(model_id, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_id}}{The model ID.}

\item{\code{config}}{Provider configuration.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoonshotLanguageModel-build_payload"></a>}}
\if{latex}{\out{\hypertarget{method-MoonshotLanguageModel-build_payload}{}}}
\subsection{Method \code{build_payload()}}{
Build the request payload for non-streaming generation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoonshotLanguageModel$build_payload(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of call options.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with url, headers, and body.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoonshotLanguageModel-build_stream_payload"></a>}}
\if{latex}{\out{\hypertarget{method-MoonshotLanguageModel-build_stream_payload}{}}}
\subsection{Method \code{build_stream_payload()}}{
Build the request payload for streaming generation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoonshotLanguageModel$build_stream_payload(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of call options.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with url, headers, and body.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoonshotLanguageModel-parse_response"></a>}}
\if{latex}{\out{\hypertarget{method-MoonshotLanguageModel-parse_response}{}}}
\subsection{Method \code{parse_response()}}{
Parse the API response into a GenerateResult.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoonshotLanguageModel$parse_response(response)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response}}{The parsed API response.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A GenerateResult object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoonshotLanguageModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MoonshotLanguageModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoonshotLanguageModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
