% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_stepfun.R
\name{StepfunImageModel}
\alias{StepfunImageModel}
\title{Stepfun Image Model Class}
\description{
Image model implementation for Stepfun image generation and editing APIs.
}
\keyword{internal}
\section{Super class}{
\code{\link[aisdk:ImageModelV1]{aisdk::ImageModelV1}} -> \code{StepfunImageModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StepfunImageModel-new}{\code{StepfunImageModel$new()}}
\item \href{#method-StepfunImageModel-do_generate_image}{\code{StepfunImageModel$do_generate_image()}}
\item \href{#method-StepfunImageModel-do_edit_image}{\code{StepfunImageModel$do_edit_image()}}
\item \href{#method-StepfunImageModel-clone}{\code{StepfunImageModel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="ImageModelV1" data-id="do_stream_image"><a href='../../aisdk/html/ImageModelV1.html#method-ImageModelV1-do_stream_image'><code>aisdk::ImageModelV1$do_stream_image()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="ImageModelV1" data-id="edit_image"><a href='../../aisdk/html/ImageModelV1.html#method-ImageModelV1-edit_image'><code>aisdk::ImageModelV1$edit_image()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="ImageModelV1" data-id="generate_image"><a href='../../aisdk/html/ImageModelV1.html#method-ImageModelV1-generate_image'><code>aisdk::ImageModelV1$generate_image()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="ImageModelV1" data-id="has_capability"><a href='../../aisdk/html/ImageModelV1.html#method-ImageModelV1-has_capability'><code>aisdk::ImageModelV1$has_capability()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="ImageModelV1" data-id="stream_image"><a href='../../aisdk/html/ImageModelV1.html#method-ImageModelV1-stream_image'><code>aisdk::ImageModelV1$stream_image()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StepfunImageModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-StepfunImageModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Stepfun image model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepfunImageModel$new(model_id, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_id}}{The model ID.}

\item{\code{config}}{Configuration list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StepfunImageModel-do_generate_image"></a>}}
\if{latex}{\out{\hypertarget{method-StepfunImageModel-do_generate_image}{}}}
\subsection{Method \code{do_generate_image()}}{
Generate images.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepfunImageModel$do_generate_image(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of call options.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A GenerateImageResult object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StepfunImageModel-do_edit_image"></a>}}
\if{latex}{\out{\hypertarget{method-StepfunImageModel-do_edit_image}{}}}
\subsection{Method \code{do_edit_image()}}{
Edit images.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepfunImageModel$do_edit_image(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of call options.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A GenerateImageResult object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StepfunImageModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StepfunImageModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepfunImageModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
