% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_stepfun.R
\name{StepfunLanguageModel}
\alias{StepfunLanguageModel}
\title{Stepfun Language Model Class}
\description{
Language model implementation for Stepfun's chat completions API.
Inherits from OpenAILanguageModel as Stepfun provides an OpenAI-compatible API.
}
\keyword{internal}
\section{Super classes}{
\code{\link[aisdk:LanguageModelV1]{aisdk::LanguageModelV1}} -> \code{\link[aisdk:OpenAILanguageModel]{aisdk::OpenAILanguageModel}} -> \code{StepfunLanguageModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StepfunLanguageModel-build_payload}{\code{StepfunLanguageModel$build_payload()}}
\item \href{#method-StepfunLanguageModel-build_stream_payload}{\code{StepfunLanguageModel$build_stream_payload()}}
\item \href{#method-StepfunLanguageModel-clone}{\code{StepfunLanguageModel$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="LanguageModelV1" data-id="generate"><a href='../../aisdk/html/LanguageModelV1.html#method-LanguageModelV1-generate'><code>aisdk::LanguageModelV1$generate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="LanguageModelV1" data-id="has_capability"><a href='../../aisdk/html/LanguageModelV1.html#method-LanguageModelV1-has_capability'><code>aisdk::LanguageModelV1$has_capability()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="LanguageModelV1" data-id="stream"><a href='../../aisdk/html/LanguageModelV1.html#method-LanguageModelV1-stream'><code>aisdk::LanguageModelV1$stream()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="do_generate"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-do_generate'><code>aisdk::OpenAILanguageModel$do_generate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="do_stream"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-do_stream'><code>aisdk::OpenAILanguageModel$do_stream()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="execute_request"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-execute_request'><code>aisdk::OpenAILanguageModel$execute_request()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="format_tool_result"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-format_tool_result'><code>aisdk::OpenAILanguageModel$format_tool_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="get_config"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-get_config'><code>aisdk::OpenAILanguageModel$get_config()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="get_history_format"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-get_history_format'><code>aisdk::OpenAILanguageModel$get_history_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="initialize"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-initialize'><code>aisdk::OpenAILanguageModel$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="parse_response"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-parse_response'><code>aisdk::OpenAILanguageModel$parse_response()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StepfunLanguageModel-build_payload"></a>}}
\if{latex}{\out{\hypertarget{method-StepfunLanguageModel-build_payload}{}}}
\subsection{Method \code{build_payload()}}{
Build the payload for the Stepfun API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepfunLanguageModel$build_payload(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of parameters for the API call.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StepfunLanguageModel-build_stream_payload"></a>}}
\if{latex}{\out{\hypertarget{method-StepfunLanguageModel-build_stream_payload}{}}}
\subsection{Method \code{build_stream_payload()}}{
Build the stream payload for the Stepfun API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepfunLanguageModel$build_stream_payload(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of parameters for the API call.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StepfunLanguageModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StepfunLanguageModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepfunLanguageModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
