% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_volcengine.R
\name{VolcengineImageModel}
\alias{VolcengineImageModel}
\title{Volcengine Image Model Class}
\description{
Image model implementation for Volcengine Ark image generation models such as
Doubao Seedream. Volcengine exposes these models through an OpenAI-compatible
image generation endpoint.
}
\keyword{internal}
\section{Super class}{
\code{\link[aisdk:ImageModelV1]{aisdk::ImageModelV1}} -> \code{VolcengineImageModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VolcengineImageModel-new}{\code{VolcengineImageModel$new()}}
\item \href{#method-VolcengineImageModel-do_generate_image}{\code{VolcengineImageModel$do_generate_image()}}
\item \href{#method-VolcengineImageModel-do_edit_image}{\code{VolcengineImageModel$do_edit_image()}}
\item \href{#method-VolcengineImageModel-clone}{\code{VolcengineImageModel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="ImageModelV1" data-id="do_stream_image"><a href='../../aisdk/html/ImageModelV1.html#method-ImageModelV1-do_stream_image'><code>aisdk::ImageModelV1$do_stream_image()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="ImageModelV1" data-id="edit_image"><a href='../../aisdk/html/ImageModelV1.html#method-ImageModelV1-edit_image'><code>aisdk::ImageModelV1$edit_image()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="ImageModelV1" data-id="generate_image"><a href='../../aisdk/html/ImageModelV1.html#method-ImageModelV1-generate_image'><code>aisdk::ImageModelV1$generate_image()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="ImageModelV1" data-id="has_capability"><a href='../../aisdk/html/ImageModelV1.html#method-ImageModelV1-has_capability'><code>aisdk::ImageModelV1$has_capability()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="ImageModelV1" data-id="stream_image"><a href='../../aisdk/html/ImageModelV1.html#method-ImageModelV1-stream_image'><code>aisdk::ImageModelV1$stream_image()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VolcengineImageModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-VolcengineImageModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Volcengine image model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VolcengineImageModel$new(model_id, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_id}}{The model ID (e.g., "doubao-seedream-5-0").}

\item{\code{config}}{Configuration list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VolcengineImageModel-do_generate_image"></a>}}
\if{latex}{\out{\hypertarget{method-VolcengineImageModel-do_generate_image}{}}}
\subsection{Method \code{do_generate_image()}}{
Generate images.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VolcengineImageModel$do_generate_image(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of call options.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A GenerateImageResult object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VolcengineImageModel-do_edit_image"></a>}}
\if{latex}{\out{\hypertarget{method-VolcengineImageModel-do_edit_image}{}}}
\subsection{Method \code{do_edit_image()}}{
Edit images by providing one or more reference images.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VolcengineImageModel$do_edit_image(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of call options.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A GenerateImageResult object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VolcengineImageModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VolcengineImageModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VolcengineImageModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
