% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_volcengine.R
\name{VolcengineLanguageModel}
\alias{VolcengineLanguageModel}
\title{Volcengine Language Model Class}
\description{
Language model implementation for Volcengine's chat completions API.
Inherits from OpenAI model but adds support for Volcengine-specific features
like reasoning content extraction from models that support \code{reasoning_content}.
}
\keyword{internal}
\section{Super classes}{
\code{\link[aisdk:LanguageModelV1]{aisdk::LanguageModelV1}} -> \code{\link[aisdk:OpenAILanguageModel]{aisdk::OpenAILanguageModel}} -> \code{VolcengineLanguageModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VolcengineLanguageModel-parse_response}{\code{VolcengineLanguageModel$parse_response()}}
\item \href{#method-VolcengineLanguageModel-clone}{\code{VolcengineLanguageModel$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="LanguageModelV1" data-id="generate"><a href='../../aisdk/html/LanguageModelV1.html#method-LanguageModelV1-generate'><code>aisdk::LanguageModelV1$generate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="LanguageModelV1" data-id="has_capability"><a href='../../aisdk/html/LanguageModelV1.html#method-LanguageModelV1-has_capability'><code>aisdk::LanguageModelV1$has_capability()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="LanguageModelV1" data-id="stream"><a href='../../aisdk/html/LanguageModelV1.html#method-LanguageModelV1-stream'><code>aisdk::LanguageModelV1$stream()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="build_payload"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-build_payload'><code>aisdk::OpenAILanguageModel$build_payload()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="build_stream_payload"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-build_stream_payload'><code>aisdk::OpenAILanguageModel$build_stream_payload()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="do_generate"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-do_generate'><code>aisdk::OpenAILanguageModel$do_generate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="do_stream"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-do_stream'><code>aisdk::OpenAILanguageModel$do_stream()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="execute_request"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-execute_request'><code>aisdk::OpenAILanguageModel$execute_request()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="format_tool_result"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-format_tool_result'><code>aisdk::OpenAILanguageModel$format_tool_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="get_config"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-get_config'><code>aisdk::OpenAILanguageModel$get_config()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="get_history_format"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-get_history_format'><code>aisdk::OpenAILanguageModel$get_history_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="initialize"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-initialize'><code>aisdk::OpenAILanguageModel$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VolcengineLanguageModel-parse_response"></a>}}
\if{latex}{\out{\hypertarget{method-VolcengineLanguageModel-parse_response}{}}}
\subsection{Method \code{parse_response()}}{
Parse the API response into a GenerateResult.
Overrides parent to extract Volcengine-specific reasoning_content.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VolcengineLanguageModel$parse_response(response)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response}}{The parsed API response.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A GenerateResult object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VolcengineLanguageModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VolcengineLanguageModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VolcengineLanguageModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
